+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- Handle multipolygon world maps
+- Handle user role of "false"
+
 ## [Version 1.2.3, Build 652] - 2019-12-20
 
 ### Fixed